app/src/test/java/com/example/max/finitestatemachiensmax/MainActivityTest.java
package com.example.max.finitestatemachiensmax;

import android.os.Build;
import android.widget.TextView;

import com.example.max.finitestatemachiensmax.Activities.MainActivity;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.robolectric.Robolectric;
import org.robolectric.RobolectricTestRunner;
import org.robolectric.annotation.Config;

import static junit.framework.Assert.assertNotNull;
import static junit.framework.Assert.assertTrue;

/**
 * Created by Max on 26.03.2017.
 */
@Config(constants = BuildConfig.class, sdk = Build.VERSION_CODES.LOLLIPOP)
@RunWith(RobolectricTestRunner.class)
public class MainActivityTest {

    private MainActivity activity;

    // @Before => JUnit 4 annotation that specifies this method should run before each test is run
    // Useful to do setup for objects that are needed in the test
    @Before
    public void setup() {
        // Convenience method to run MainActivity through the Activity Lifecycle methods:
        // onCreate(...) => onStart() => onPostCreate(...) => onResume()
        activity = Robolectric.setupActivity(MainActivity.class);
    }

    // @Test => JUnit 4 annotation specifying this is a test to be run
    // The test simply checks that our TextView exists and has the text "Hello world!"
    @Test
    public void validateTextViewContent() {
        TextView mainText = (TextView) activity.findViewById(R.id.state_text_view);
        assertNotNull("TextView could not be found", mainText);
        assertTrue("TextView contains incorrect text",
                "Test Card IO and Stripe".equals(mainText.getText().toString()));
    }
}
